Orgnaise code into packages
