+++ src/main/java/cheryl/command/AddDeadlineCommand.java
@@ -1,3 +1,11 @@
+package cheryl.command;
+
+import cheryl.task.Task;
+import cheryl.task.Deadline;
+import cheryl.util.DukeException;
+import cheryl.util.Storage;
+import cheryl.util.TaskList;
+import cheryl.util.Ui;
 import java.time.LocalDate;
 import java.time.format.DateTimeParseException;
