+++ src/main/java/cheryl/util/DukeException.java
@@ -1,3 +1,5 @@
+package cheryl.util;
+
 public class DukeException extends Exception {
     public DukeException(String message) {
         super(message);
